C++: Finished the Market Data Subscription code.    Yay! That completes my C++ version which means C++, C#, and Java are complete. C# & Java: I moved most of the display format code to a DisplayFormat type to help with any internationalization issues that may crop up.    I still have to do this for C++. Java: Session code uses this._marketSimulatorTypeClass.start() instead of this._marketSimulatorTypeClass.run() so that it starts a new thread.    - Thanks to Codeplex user danhnguyen for pointing this out. Java: com.bemu.BEmu.types.randomDouble now uses a better way to round doubles to four digits. Java & C#: I now implement a Session.NextEvent(long) function that exists from the actual API which accepts a timeout in milleseconds.    - Thanks to Codeplex user pichulines for pointing this out.
Former-commit-id: 6908585d54fdba039266990ce96b8bd0ebab59a8

+++ BEmu_cpp/src/ReferenceDataRequest/ElementReferenceArrayChainTickers.cpp
@@ -28,7 +28,7 @@ namespace BEmu
 			{
 				dtExp = Datetime::Today();
 				dtExp.addMonths(1);
-				dtExp.addDays(-dtExp.day());
+				dtExp.addDays(-((long)dtExp.day()));
 				dtExp.addDays(20); //assume the 20th of the month
 			}
 			else if(strDtExp.length() == 8)
@@ -46,14 +46,14 @@ namespace BEmu
 
 				dtExp = Datetime::Today();
 				dtExp.addMonths(numMonths);
-				dtExp.addDays(-dtExp.day());
+				dtExp.addDays(-((long)dtExp.day()));
 				dtExp.addDays(20); //assume the 20th of the month
 			}
 			else
 				throw elementPtrEx;
 
 			int strike = RandomDataGenerator::Strike();
-			for (int count = 0; count < numPoints; count++, strike += 5)
+			for (unsigned count = 0; count < numPoints; count++, strike += 5)
 			{
 				ElementPtr * elm = new ElementReferenceArrayChainTickersItem(ticker, dtExp, optionality, strike);
 				this->_values.push_back(elm);
